feat: add Jacoco for code coverage and implement BlockHound for blocking call detection in tests
