+++ src/integrationTest/java/com/code/agent/github/adapter/GitHubAdapterTest.java
@@ -1,7 +1,8 @@
-package com.code.agent.infra.github.adapter;
+package com.code.agent.github.adapter;
 
 import com.code.agent.domain.model.PullRequestReviewInfo;
 import com.code.agent.infra.config.GitHubProperties;
+import com.code.agent.infra.github.adapter.GitHubAdapter;
 import com.code.agent.infra.github.util.GitHubRetryUtil;
 import okhttp3.mockwebserver.MockResponse;
 import okhttp3.mockwebserver.MockWebServer;
